refactor: implement `TryFrom` trait for Property->ParamType conversion
