+++ packages/fuels-core/src/code_gen/custom_types_gen.rs
@@ -1,4 +1,3 @@
-use crate::parse::parse_param_type_from_property;
 use crate::types::expand_type;
 use crate::utils::{has_array_format, ident};
 use crate::ParamType;
@@ -38,7 +37,7 @@ pub fn expand_custom_struct(prop: &Property) -> Result<TokenStream, Error> {
     // 2. The creation of a token and its insertion into a vector of Tokens.
     for component in components {
         let field_name = ident(&component.name.to_snake_case());
-        let param_type = parse_param_type_from_property(component)?;
+        let param_type = ParamType::try_from(component)?;
 
         match param_type {
             // Case where a struct takes another struct
@@ -196,7 +195,7 @@ pub fn expand_custom_enum(enum_name: &str, prop: &Property) -> Result<TokenStrea
         let variant_name = ident(&component.name);
         let dis = discriminant as u8;
 
-        let param_type = parse_param_type_from_property(component)?;
+        let param_type = ParamType::try_from(component)?;
         match param_type {
             // Case where an enum takes another enum
             ParamType::Enum(_params) => {
