+++ packages/fuels-core/src/code_gen/functions_gen.rs
@@ -2,7 +2,6 @@ use crate::abi_encoder::ABIEncoder;
 use crate::code_gen::custom_types_gen::extract_custom_type_name_from_abi_property;
 use crate::code_gen::docs_gen::expand_doc;
 use crate::json_abi::ABIParser;
-use crate::parse::parse_param_type_from_property;
 use crate::types::expand_type;
 use crate::utils::{ident, safe_ident};
 use crate::{ParamType, Selector};
@@ -55,7 +54,7 @@ pub fn expand_function(
     // be used to be tokenized and passed onto `method_hash()`.
     let output_param = match &function.outputs[..] {
         [output] => {
-            let param_type = parse_param_type_from_property(output).unwrap();
+            let param_type = ParamType::try_from(output).unwrap();
 
             let tok: proc_macro2::TokenStream =
                 format!("Some(ParamType::{})", param_type).parse().unwrap();
@@ -89,7 +88,7 @@ fn expand_fn_outputs(outputs: &[Property]) -> Result<TokenStream, Error> {
         [output] => {
             // If it's a primitive type, simply parse and expand.
             if !output.is_custom_type() {
-                return expand_type(&parse_param_type_from_property(output)?);
+                return expand_type(&ParamType::try_from(output)?);
             }
 
             // If it's a {struct, enum} as the type of a function's output, use its tokenized name only.
@@ -223,7 +222,7 @@ fn expand_function_arguments(
         };
 
         // TokenStream representing the type of the argument
-        let kind = parse_param_type_from_property(param)?;
+        let kind = ParamType::try_from(param)?;
 
         // If it's a tuple, don't expand it, just use the type signature as it is (minus the string "struct " | "enum ").
         let tok = if let ParamType::Tuple(_tuple) = kind {
@@ -235,7 +234,7 @@ fn expand_function_arguments(
             expand_input_param(
                 fun,
                 &param.name,
-                &parse_param_type_from_property(param)?,
+                &ParamType::try_from(param)?,
                 &custom_property,
             )?
         };
@@ -272,7 +271,7 @@ fn build_expanded_tuple_params(tuple_param: &Property) -> Result<String, Error> 
         .expect("tuple parameter should have components")
     {
         if !component.is_custom_type() {
-            let p = parse_param_type_from_property(component)?;
+            let p = ParamType::try_from(component)?;
             let tok = expand_type(&p)?;
             toks.push_str(&tok.to_string());
         } else {
